refactor: rename memberType to avoid clashing with Zeebe classes
